fix: add room service interface
