+++ src/main/java/com/trianglechoke/codesparring/room/service/RoomMemberService.java
@@ -1,11 +1,3 @@
 package com.trianglechoke.codesparring.room.service;
 
-import com.trianglechoke.codesparring.room.dto.RoomMemberDTO;
-
-public class RoomMemberService {
-    /* 대기방에 회원 추가 */
-    public void addRoomMember(RoomMemberDTO roomMemberDTO) {}
-
-    /* 대기방에 회원 삭제 */
-    public void removeMember(RoomMemberDTO roomMemberDTO) {}
-}
+public interface RoomMemberService {}
