+++ src/main/java/com/trianglechoke/codesparring/room/service/RoomMemberServiceImpl.java
@@ -1,0 +1,11 @@
+package com.trianglechoke.codesparring.room.service;
+
+import com.trianglechoke.codesparring.room.dto.RoomMemberDTO;
+
+public class RoomMemberServiceImpl {
+    /* 대기방에 회원 추가 */
+    public void addRoomMember(RoomMemberDTO roomMemberDTO) {}
+
+    /* 대기방에 회원 삭제 */
+    public void removeMember(RoomMemberDTO roomMemberDTO) {}
+}
